test(onboarding): share authenticated boundary
