+++ packages/backend/test/onboarding.ts
@@ -2,13 +2,42 @@ import {
   LearningProgramKeySchema,
   LearningProgramSchema,
 } from "@nakafa/aksara-contracts/program/spec";
-import type { createConvexTestWithBetterAuth } from "@repo/backend/convex/test.helpers";
+import type { Doc } from "@repo/backend/convex/_generated/dataModel";
+import {
+  createConvexTestWithBetterAuth,
+  seedAuthenticatedUser,
+} from "@repo/backend/convex/test.helpers";
 import {
   activateProgramSnapshot,
   makeProgramSnapshotData,
   makeTechnicalProgram,
 } from "@repo/backend/test/program/snapshot";
 import { Effect } from "effect";
+
+/** Creates one isolated authenticated boundary for onboarding behavior. */
+export const createOnboardingTest = Effect.fn("test.onboarding.create")(
+  function* (now: number, role?: Doc<"users">["role"], suffix = "onboarding") {
+    yield* Effect.sync(() => vi.setSystemTime(now));
+    const test = createConvexTestWithBetterAuth();
+    const identity = yield* Effect.promise(() =>
+      test.mutation((ctx) =>
+        seedAuthenticatedUser(ctx, {
+          now,
+          suffix,
+          ...(role === undefined ? {} : { role }),
+        })
+      )
+    );
+    return {
+      authenticated: test.withIdentity({
+        sessionId: identity.sessionId,
+        subject: identity.authUserId,
+      }),
+      identity,
+      test,
+    };
+  }
+);
 
 /** Activates every curriculum key used by onboarding defaults. */
 export const activateOnboardingPrograms = Effect.fn(
